add ability to have arguments prefix flags
normally all flags must come before any arguments. this allows arguments
to come before and after the flags. allowing both isn't really the end
goal but is allowed. this is a feature in case for where people have
commands that makes more sense to have the argument (or 'target') as
the argument before any modifying flags for said target